output/kaggle-upload/supermix-source/source/requirements_train_build.txt
torch
pillow
nltk

# Optional acceleration backends (install only if needed)
# torch-directml

